RequestParser를 주입하는 것에서 static으로 변경
